Switch to L5K format.
This removes the hash and exclusions report sections. The former because
L5K is always exported, so the resulting hash isn't traceable to the
original source(ACD). The latter because AOI local tags can now be
evaluated and L5K doesn't omit structured tag data like L5X, so no
exclusions are necessary.